Volca Drum Patch values aligned better

+++ src/devices/volca_drum/volca_drum_patch.rs
@@ -12,12 +12,12 @@ pub struct VolcaDrumPatchLayout {
     pub sound_src_type: VolcaDrumPatchLayoutSoundSrcType,
     pub mod_type: VolcaDrumPatchLayoutModType,
     pub amp_eg: VolcaDrumPatchLayoutAmpEg,
-    pub level: usize,
-    pub pitch: usize,
-    pub eg_attack: usize,
-    pub eg_release: usize,
-    pub mod_amount: usize,
-    pub mod_rate: usize,
+    pub level: u8,      // From 0 to 255.
+    pub pitch: u8,      // From 0 to 255.
+    pub eg_attack: u8,  // From 0 to 255.
+    pub eg_release: u8, // From 0 to 255.
+    pub mod_amount: u8, // TODO: Should go from -100 to +100. For now 128=0.
+    pub mod_rate: u8,   // From 0 to 255.
 }
 #[derive(Clone, Debug, PartialEq)]
 pub enum VolcaDrumPatchLayoutSoundSrcType {
